feat: implement bst_inorder
I broke here with the Java template, and instead of producing a list
decided to accept a callback function.
If the user wants to produce a list, the user can do so by passing a
callback that will collect entries into the desired list.
I provided a simpler example `print_entry` that will simply print the
entries to stdout.

Otherwise, this is essentially a one-to-one copy of the
Java-implementation algorithm, it's quite convoluted and I'm sure this
could be simplified a lot.
Maybe in a future commit...