fix(avl): stop retracing when a bf incs/decs to 0
Since a nodes bf is calculated with the formula bf = height(right) -
height(left) and because height is always the hight of the highest
subtree, if the bf reaches 0, it means that the other subtree was one
larger, now they are both the same height. That means, that the overall
height didn't change at all. Therefore, the parent's bf does not change
either.